refactor(components): move ProductCard to common components
