+++ src/components/users/home/PopularProduct.jsx
@@ -1,4 +1,4 @@
-import ProductCard from "./ProductCard";
+import ProductCard from "../../common/ProductCard";
 
 export default function PopularProduct() {
   const products = [
@@ -66,7 +66,6 @@ export default function PopularProduct() {
         Produk Populer
       </h1>
 
-      {/* Layout responsif menggunakan grid */}
       <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-6 mt-6">
         {products.map((product, index) => (
           <ProductCard
